+++ nens_auth_client/models.py
@@ -1,4 +1,6 @@
 # (c) Nelen & Schuurmans.  Proprietary, see LICENSE file.
+from .signals import invitation_accepted
+
 from django.conf import settings
 from django.core.mail import send_mail
 from django.core.exceptions import PermissionDenied
@@ -36,7 +38,22 @@ class RemoteUser(models.Model):
         help_text="The user ID in the external identity provider, which is present as the 'sub' field in tokens.",
     )
     created = models.DateTimeField(auto_now_add=True)
-    last_modified = models.DateTimeField(auto_now=True)
+    last_modified = models.DateTimeField(
+        auto_now=True,
+        help_text="The last time this remote user logged in.",
+    )
+    id_token = models.TextField(
+        blank=True,
+        help_text="The most recent ID token provided by the external identity provider.",
+    )
+    access_token = models.TextField(
+        blank=True,
+        help_text="The most recent access token provided by the external identity provider.",
+    )
+    refresh_token = models.TextField(
+        blank=True,
+        help_text="The most recent refresh token provided by the external identity provider.",
+    )
 
     def __str__(self):
         return self.external_user_id
@@ -118,7 +135,9 @@ class Invitation(models.Model):
 
     @property
     def expires_at(self):
-        return self.created_at + timedelta(days=settings.NENS_AUTH_INVITATION_EXPIRY_DAYS)
+        return self.created_at + timedelta(
+            days=settings.NENS_AUTH_INVITATION_EXPIRY_DAYS
+        )
 
     def check_acceptability(self):
         """Checks if this invitation is PENDING and if it has not expired
@@ -144,6 +163,9 @@ class Invitation(models.Model):
             raise
         else:
             self._update_status(Invitation.ACCEPTED)
+            invitation_accepted.send(
+                sender=self.__class__, obj=self, user=user
+            )
             return result
 
     def reject(self):